requirements.txt
pygame==1.9.6
pyopengl==3.1.5

